+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 ### Fixed
 - (maint) Fixed a bug in the 'artifact already exists' error where the path to the artifact
   wasn't printing correctly
+- (maint) Go back to old ERB.new method call to accommodate Ruby 2.5
 
 ## [0.109.0] - 2023-02-28
 ### Changed