+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased]
+
+## [0.109.1] - 2023-03-01
 ### Fixed
 - (maint) Fixed a bug in the 'artifact already exists' error where the path to the artifact
   wasn't printing correctly
@@ -881,7 +883,8 @@ this is a final version.
 
 ## Versions <= 0.5.0 do not have a change log entry
 
-[Unreleased]: https://github.com/puppetlabs/packaging/compare/0.109.0...HEAD
+[Unreleased]: https://github.com/puppetlabs/packaging/compare/0.109.1...HEAD
+[0.109.1]: https://github.com/puppetlabs/packaging/compare/0.109.0...0.109.1
 [0.109.0]: https://github.com/puppetlabs/packaging/compare/0.108.2...0.109.0
 [0.108.2]: https://github.com/puppetlabs/packaging/compare/0.108.1...0.108.2
 [0.108.1]: https://github.com/puppetlabs/packaging/compare/0.108.0...0.108.1